==> 01_01_asking-questions.Rmd <==
# The scientific process {#scientific-process}

> *There is something fascinating about science. One gets such a wholesale return of conjecture out of a trifling investment of fact.*
>
> Mark Twain
>
> *To do science is to search for patterns, not simply to accumulate facts.*
>
> Robert MacArthur

## Stages in the scientific process {#stages-scientific-process}

Science is about asking, and answering, the right questions. A number of distinct stages occur within this process: making observations, asking questions, formulating hypotheses, and testing predictions. Collectively these are the building blocks of what is known as the scientific method. Exactly how they fit together, and what the philosophical and practical limitations of different approaches are, have been the subject of much debate by philosophers of science. We're not going to tackle those issues here, but instead try to extract a general working framework for the process of a typical scientific investigation.

### Observations {#stages-observations}

**Observation** --- *information, or impression, about events or objects*.

In general the questions we ask are not generated by pure abstract thought, but are a result of observations about the natural world. These may take the form of direct observations we make ourselves, patterns that crop up in data collected for other purposes, in non-specific surveys, and the accumulated works of other people.

Such observations of biological systems will lead almost automatically into asking questions.

### Questions {#stages-questions}

**Question** --- *what it is that you want to know; the scope of your investigation*.

It is important to keep the question sufficiently focussed. The overall aim of a particular study should be to answer one or (perhaps) two questions. The question of why the tropics are more diverse than the temperate regions is a valid and fascinating question, but it's probably not be a good choice for a final year project or even a PhD.

The next stage is to formulate an hypothesis.

### Hypotheses {#stages-hypotheses}

**Hypothesis** --- *an explanation proposed to account for observed facts.*

In general, in biology, the critical distinguishing feature of an hypothesis is that it specifies some biological process, or processes, which might account for the observations made. Formulating hypotheses requires more than just a restatement of the question -- it usually embodies some mechanism (though this may not be fully understood) and it will often draw on additional information. One question will also often generate more than one hypothesis.

Once we have formulated a set of hypotheses, we need to derive the predictions that follow.

### Predictions {#stages-predictions}

**Prediction** --- *what you would expect to see if the hypothesis was true*.

Hypotheses are about proposing explanations, but they might not be directly testable. That is, they may not tell us what data to collect, or what pattern to expect in the data. To be able to test an hypothesis we need to make some predictions from that hypothesis. These will be determined both by what we expect to see and what it is possible, or practical, to measure. A prediction is not simply a rephrasing of the hypothesis -- it should more or less lead to a statement of the experiment to conduct or observation to make, and type of data to collect.

The ideal prediction is one that is unique to the hypothesis it is based on, so if the prediction is true only one of the hypotheses could have been responsible. It may not always be possible to generate such 'clean' predictions, in which case, some combination of predictions may need to be considered. Additionally, several processes may be operating at the same time, which makes hypothesis testing harder still, because it may be necessary to consider two or more hypotheses, and their corresponding predictions, in combination.

Science is hard!

## An example

#### Observations {.unnumbered}

Imagine, while observing a stream one day, we notice that the freshwater 'shrimps' (*Gammarus*) seem to occur almost entirely under stones; we rarely seem to see them in a patch of open stream bed. Having made observations, it may be necessary to collect some more data to check that this phenomenon is not just a one-off event, or a false impression. Look under a few more stones, watch the same species another day, or in another place, check the literature for similar observations by others.

#### Question {.unnumbered}

Why does *Gammarus* spend most of its time under stones?

#### Hypotheses {.unnumbered}

*Gammarus* occur under stones because:

-   they need to shelter from the current

-   their food (leaf litter) gets trapped and accumulates under stones

-   they are subject to predation by visually hunting fish and need to remain out of sight

#### Predictions {.unnumbered}

Taking each hypothesis in turn, we expect to observe:

-   *Shelter hypothesis:* a greater proportion of *Gammarus* should be found in the open in streams with slow flow, or in slower flowing areas of a stream.

-   *Food hypothesis:* *Gammarus* should not aggregate under stones from which all leaf litter has been removed; *Gammarus* should aggregate on patches of leaf litter tethered in the open part of the stream bed.

-   *Predation hypothesis:* *Gammarus* should aggregate under stones more in streams where fish are present than where they are not; *Gammarus* may spend less time under stones at night.

If we want to tease apart our hypotheses we may need to refine these predictions. For example, *Gammarus* may be under stones because it prefers the sheltered environment, but also because food accumulates there. In this case we might expect that *Gammarus* will show a weak aggregative response to shelter alone, or food alone, and a stronger one to them both together.

## Hypothesis testing

Once we have firmed up our hypotheses and predictions we are in the position to collect data to evaluate our ideas. On the basis of these data we will either accept or reject our various hypotheses. The important thing to realise about the process of hypothesis testing is that, in science especially, hypotheses are either rejected, or not rejected, but an hypothesis can rarely, except in trivial cases, be proved.

Why is this? Since we cannot be sure that we've thought of all the possible hypotheses to explain an observation, finding evidence that supports a prediction does not guarantee that the underlying hypothesis is the only one which could have produced the effect we find. On the other hand, if we find evidence that directly contradicts the prediction(s) from an hypothesis, we can be certain (assuming the prediction and data are not flawed) that the hypothesis cannot be true.

An hypothesis which predicted that all conifers should be evergreen could be supported by numerous observations of different conifer species in forests around the world, but is conclusively refuted by the first larch tree we encounter.

Having tested our hypothesis, by examining the evidence that its predictions are true, we may accept it as the best current explanation of the observations, or we may reject it as an explanation, and turn to other hypotheses. The same procedure must then be repeated for these hypotheses.

This basic cycle of proposing hypotheses and then seeking evidence potentially capable of falsifying them, is, in essence, the idealized model of the scientific process famously proposed by the philosopher of science Karl Popper (1902-1994). It is often termed *falsificationism*.

## Don't we ever know anything for sure? {#are-we-sure}

The method presented here provides a view of science as one in which we suggest hypotheses, then test them, trying to reject them by finding conclusive counter-evidence, then replacing them with new hypotheses. It all sounds a bit frustrating.

In fact of course we do 'accept' hypotheses all the time -- that is, we fail to reject them over and over again. These hypotheses become more accepted and in some sense become regarded as 'true' if repeated attempts to test them all fail to provide good counter evidence. In other words, we have some ideas that are doing pretty well in terms of resisting falsification, and we use these as our best estimates of the truth, with the proviso that it is still possible a better idea will come along in due course.

The simple process of falsification described above also presents a picture of scientists as neutral, objective creatures, rationally proceeding through cycles of setting up hypotheses, testing them, rejecting them, setting them aside and starting over again. Of course this is not a true reflection of the complex, often messy, business involved in trying to figure out how the world works. Philosophers of science have argued long and hard about how far from this idealized process real science actually is. Various alternative philosophies suggest more 'realistic' processes. For example...

-   Thomas Kuhn's view of science as periods of relative stasis, where people work within an accepted paradigm (a set of views about how things work) despite accumulating evidence that doesn't always support the paradigm, until finally it is upset by a 'revolution' which rejects the entire paradigm, and proposes a new view. This is where the phrase 'paradigm shift' originated.

-   Imre Lakatos proposed some resolution of Kuhn's views, suggesting that scientific ideas were grouped together in 'research programmes' concerned with particular endeavours, and that within these there may be core ideas that are not challenged, but other related ideas that are being challenged and adjusted by falsification, and that together these make each research programme progress.

This is a very over-simplified sketch of some important ideas. These are well worth reading about, but in practice most philosophical arguments are more focused on how whole areas of science develop. When thinking about how to construct a study of one problem, the basic falsification cycle is a pretty good approach to have in mind. Keep in mind that the process laid out here is not a strict set of rules, but outlines an approach to scientific investigation which is widely considered to provide a rigorous and productive system. As with all such systems understanding the 'normal' process is a prerequisite for constructively breaking the rules.
